+++ src/components/Editor/EditorSlice.js
@@ -3,11 +3,11 @@ import { createOrUpdateProject, readProject, createRemix } from '../../utils/api
 
 export const syncProject = (actionName) => createAsyncThunk(
   `editor/${actionName}Project`,
-  async({ project, identifier, accessToken, autosave }, { rejectWithValue }) => {
+  async({ project, identifier, projectType, accessToken, autosave }, { rejectWithValue }) => {
     let response
     switch(actionName) {
       case 'load':
-        response = await readProject(identifier, accessToken)
+        response = await readProject(identifier, projectType, accessToken)
         break
       case 'remix':
         response = await createRemix(project, accessToken)
@@ -65,8 +65,11 @@ export const EditorSlice = createSlice({
     lastSavedTime: null,
     senseHatAlwaysEnabled: false,
     senseHatEnabled: false,
+    accessDeniedNoAuthModalShowing: false,
+    accessDeniedWithAuthModalShowing: false,
     betaModalShowing: false,
     loginToSaveModalShowing: false,
+    notFoundModalShowing: false,
     renameFileModalShowing: false,
     modals: {},
   },
@@ -167,6 +170,13 @@ export const EditorSlice = createSlice({
     setProjectListLoaded: (state, action) => {
       state.projectListLoaded = action.payload;
     },
+    closeAccessDeniedNoAuthModal: (state) => {
+      state.accessDeniedNoAuthModalShowing = false
+      state.modals = {}
+    },
+    closeAccessDeniedWithAuthModal: (state) => {
+      state.accessDeniedWithAuthModalShowing = false
+    },
     showBetaModal: (state) => {
       state.betaModalShowing = true
     },
@@ -178,6 +188,9 @@ export const EditorSlice = createSlice({
     },
     closeLoginToSaveModal: (state) => {
       state.loginToSaveModalShowing = false
+    },
+    closeNotFoundModal: (state) => {
+      state.notFoundModalShowing = false
     },
     showRenameFileModal: (state, action) => {
       state.modals.renameFile = action.payload
@@ -215,6 +228,8 @@ export const EditorSlice = createSlice({
     })
     builder.addCase('editor/loadProject/pending', (state, action) => {
       state.loading = 'pending'
+      state.accessDeniedNoAuthModalShowing = false
+      state.modals = {}
       state.currentLoadingRequestId = action.meta.requestId
     })
     builder.addCase('editor/loadProject/fulfilled', (state, action) => {
@@ -226,10 +241,24 @@ export const EditorSlice = createSlice({
         state.currentLoadingRequestId = undefined
       }
     })
+
     builder.addCase('editor/loadProject/rejected', (state, action) => {
       if (state.loading === 'pending' && state.currentLoadingRequestId === action.meta.requestId) {
         state.loading = 'failed'
         state.saving = 'idle'
+        const splitErrorMessage = action.error.message.split(' ')
+        const errorCode = splitErrorMessage[splitErrorMessage.length - 1]
+        if (errorCode === '404') {
+          state.notFoundModalShowing = true
+        } else if ((errorCode === '500' || errorCode === '403') && action.meta.arg.accessToken) {
+          state.accessDeniedWithAuthModalShowing = true
+        } else if ((errorCode === '500' || errorCode === '403') && !action.meta.arg.accessToken) {
+          state.accessDeniedNoAuthModalShowing = true
+          state.modals.accessDenied = {
+            identifier: action.meta.arg.identifier,
+            projectType: action.meta.arg.projectType
+          }
+        }
         state.currentLoadingRequestId = undefined
       }
     })
@@ -261,10 +290,13 @@ export const {
   updateImages,
   updateProjectComponent,
   updateProjectName,
+  closeAccessDeniedNoAuthModal,
+  closeAccessDeniedWithAuthModal,
   showBetaModal,
   closeBetaModal,
   showLoginToSaveModal,
   closeLoginToSaveModal,
+  closeNotFoundModal,
   showRenameFileModal,
   closeRenameFileModal,
 } = EditorSlice.actions