+++ CHANGELOG.md
@@ -8,8 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Plausbile event tracking for login, remix, save and code run (#250)
-- Message prompting users to login or save if they make non-autosaved changes (#298)
-- Unit tests for the autosave trigger (#298)
+- Message prompting users to login or save if they make non-autosaved changes (#291)
+- Unit tests for the autosave trigger (#291)
+- Project not found and access denied modals shown on project loading error (#298)
 
 ## [0.9.0]